requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
pyodbc>=4.0.0
streamlit>=1.28.0
python-dotenv>=0.19.0
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0
fastapi>=0.100.0
uvicorn>=0.23.0 